pyproject.toml
[tool.poetry]
name = "simweb"
version = "0.1.0"
description = ""
authors = ["Yunier Rojas García <yunier.rojas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
simpy = "^4.1.1"
pandas = "^2.3.2"
plotly = "^6.3.0"
pyyaml = "^6.0.2"
polars = "^1.33.1"
pyarrow = "^21.0.0"
tqdm = "^4.67.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
